Split preferences from __init__.py to preference.py file.
